data_lake/requirements.txt
# Core Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
pyarrow>=12.0.0,<16.0.0
polars>=0.17.0,<1.0.0

# Data Storage & Access
boto3>=1.26.0,<2.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.3,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Data Quality & Validation
great-expectations>=0.16.0,<1.0.0
pandera>=0.14.0,<1.0.0

# Workflow & Orchestration
prefect>=2.0.0,<3.0.0

# Data Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<0.14.0
plotly>=5.13.0,<6.0.0

# Machine Learning
scikit-learn>=1.2.0,<2.0.0
torch>=2.0.0,<3.0.0
transformers>=4.25.0,<5.0.0

# Utilities
tqdm>=4.65.0,<5.0.0
loguru>=0.6.0,<1.0.0
python-dateutil>=2.8.2,<3.0.0
pytz>=2022.7,<2026.0

# Testing
pytest>=7.3.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0

# Documentation
sphinx>=6.0.0,<7.0.0
sphinx-rtd-theme>=1.2.0,<2.0.0

# API Clients
requests>=2.28.0,<3.0.0
fastapi>=0.95.0,<1.0.0
uvicorn>=0.21.0,<1.0.0
facebook-business>=17.0.0,<18.0.0

# Data Exploration & Parsing
beautifulsoup4>=4.11.0,<5.0.0
lxml>=4.9.0,<5.0.0
matplotlib>=3.7.0,<4.0.0

# Development Tools
jupyter>=1.0.0,<2.0.0
ipython>=8.0.0,<9.0.0
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.0.0,<2.0.0

# Version Control
gitpython>=3.1.0,<4.0.0

# Containerization
docker>=6.0.0,<7.0.0

# For data lake flow visualization
graphviz>=0.20.0,<0.21.0
pydot>=1.4.2,<2.0.0

statsmodels